social-assets/pyproject.toml
[project]
name = "zech-sh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.2",
    "pillow>=11.3.0",
]
